♻️ Move localization-strings under localization service (#34945)
* Make #services/localization a submodule

* Move localized-strings closer to localization service

* Update owners

* Drop index exports for cleaner deps

* Update imports

* Update dep-check config
