+++ build-system/test-configs/dep-check-config.js
@@ -425,8 +425,10 @@ exports.rules = [
       'extensions/amp-link-rewriter/0.1/amp-link-rewriter.js->' +
         'src/service/navigation.js',
       // For localization.
-      'extensions/amp-story/1.0/amp-story-localization-service.js->src/service/localization.js',
-      'extensions/amp-story-auto-ads/0.1/story-ad-localization.js->src/service/localization.js',
+      'extensions/amp-story/1.0/amp-story-localization-service.js->src/service/localization/index.js',
+      'extensions/amp-story*/**/*.js->src/service/localization/strings.js',
+      'extensions/amp-story-auto-ads/0.1/story-ad-localization.js->src/service/localization/index.js',
+
       // Accessing calculateScriptBaseUrl() for vendor config URLs
       'extensions/amp-analytics/0.1/config.js->' +
         'src/service/extension-script.js',
