+++ extensions/amp-story-360/0.1/amp-story-360.js
@@ -25,7 +25,7 @@ import {
 } from '../../amp-story/1.0/amp-story-store-service';
 import {CSS} from '../../../build/amp-story-360-0.1.css';
 import {CommonSignals} from '#core/constants/common-signals';
-import {LocalizedStringId} from '../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import {Matrix, Renderer} from '#third_party/zuho/zuho';
 import {Services} from '#service';
 import {applyFillContent, isLayoutSizeDefined} from '#core/dom/layout';
