+++ extensions/amp-story-auto-ads/0.1/test/test-story-ad-localization-strings.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {LocalizedStringId} from '../../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import localesObjs from '../_locales';
 
 describes.fakeWin('amp-story-ad-localization-strings', {amp: true}, () => {
