+++ extensions/amp-story/1.0/amp-story-draggable-drawer.js
@@ -22,7 +22,7 @@ import {
 } from './amp-story-store-service';
 import {CSS} from '../../../build/amp-story-draggable-drawer-header-1.0.css';
 import {Layout} from '#core/dom/layout';
-import {LocalizedStringId} from '../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import {Services} from '#service';
 import {closest} from '#core/dom/query';
 import {createShadowRootWithStyle} from './utils';
