+++ extensions/amp-story/1.0/amp-story-embedded-component.js
@@ -30,7 +30,7 @@ import {
 import {CSS} from '../../../build/amp-story-tooltip-1.0.css';
 import {EventType, dispatch} from './events';
 import {Keys} from '#core/constants/key-codes';
-import {LocalizedStringId} from '../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import {Services} from '#service';
 import {addAttributesToElement, tryFocus} from '#core/dom';
 import {closest, matches} from '#core/dom/query';
