+++ extensions/amp-story/1.0/amp-story-page.js
@@ -48,7 +48,7 @@ import {Deferred} from '#core/data-structures/promise';
 import {EventType, dispatch} from './events';
 import {Layout} from '#core/dom/layout';
 import {LoadingSpinner} from './loading-spinner';
-import {LocalizedStringId} from '../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import {MediaPool} from './media-pool';
 import {Services} from '#service';
 import {StoryAdSegmentTimes} from '#experiments/story-ad-progress-segment';
