+++ extensions/amp-story/1.0/amp-story-system-layer.js
@@ -26,7 +26,7 @@ import {
   DevelopmentModeLog,
   DevelopmentModeLogButtonSet,
 } from './development-ui';
-import {LocalizedStringId} from '../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import {ProgressBar} from './progress-bar';
 import {Services} from '#service';
 import {closest, matches, scopedQuerySelector} from '#core/dom/query';
