+++ extensions/amp-story/1.0/amp-story-unsupported-browser-layer.js
@@ -16,7 +16,7 @@
 
 import {Action, getStoreService} from './amp-story-store-service';
 import {CSS} from '../../../build/amp-story-unsupported-browser-layer-1.0.css';
-import {LocalizedStringId} from '../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import {createShadowRootWithStyle} from './utils';
 import {dict} from '#core/types/object';
 import {removeElement} from '#core/dom';
