+++ extensions/amp-story/1.0/test/test-amp-story-localization-strings.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {LocalizedStringId} from '../../../../src/localized-strings';
+import {LocalizedStringId} from '#service/localization/strings';
 import localesObjs from '../_locales';
 
 describes.fakeWin('amp-story-localization-strings', {amp: true}, () => {
