+++ src/service/localization/OWNERS
@@ -1,0 +1,32 @@
+// For an explanation of the OWNERS rules and syntax, see:
+// https://github.com/ampproject/amp-github-apps/blob/main/owners/OWNERS.example
+
+{
+  rules: [
+    {
+      pattern: 'strings.js',
+      owners: [
+        {
+          name: 'gmajoulet',
+          notify: true,
+        },
+        {
+          name: 'newmuis',
+          notify: true,
+        },
+        {
+          name: 'enriqe',
+        },
+        {
+          name: 'processprocess',
+        },
+        {
+          name: 'mszylkowski',
+        },
+        {
+          name: 'raxsha',
+        },
+      ],
+    },
+  ],
+}
