+++ test/unit/test-localization.js
@@ -21,7 +21,8 @@ import {
 import {
   LocalizedStringId,
   createPseudoLocale,
-} from '../../src/localized-strings';
+} from '#service/localization/strings';
+
 import {Services} from '#service';
 
 describes.fakeWin('localization', {amp: true}, (env) => {
